+++ packages/common/src/core/traits/IWithAssistant.ts
@@ -1,0 +1,11 @@
+import { IHubAssistant } from "../../assistants";
+
+/**
+ * Expose an assistant on an entity
+ */
+export interface IWithAssistant {
+  /**
+   * Optional Hub Assistant.
+   */
+  assistant?: IHubAssistant;
+}